add test to walletServiceImpl

+++ src/main/java/com/michael/walletapi/service/Impl/TransactionServiceImpl.java
@@ -25,7 +25,7 @@ public class TransactionServiceImpl implements TransactionService {
     }
 
     public Transaction createTopupTransaction(Wallet existWallet, TransactionDTO transactionDTO){
-        TransactionType existType = this.transactionTypeService.getTransactionTypeById((long) transactionDTO.getTransaction_type_id());
+        TransactionType existType = this.transactionTypeService.getTransactionTypeById(transactionDTO.getTransaction_type_id());
 
         if(existType == null){
             return null;
